Renamed OpenNeuroBase dataset class to OpenNEuroDataset
