+++ brainsets/datasets/KlinzingSleepDS005555.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional
 
-from brainsets.datasets.OpenNeuroBase import OpenNeuroDataset, OpenNeuroSplitType
+from brainsets.datasets.OpenNeuroDataset import OpenNeuroDataset, OpenNeuroSplitType
 
 
 class KlinzingSleepDS005555(OpenNeuroDataset):
