+++ brainsets/datasets/KochiVisualNamingDS006914.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional
 
-from brainsets.datasets.OpenNeuroBase import OpenNeuroDataset, OpenNeuroSplitType
+from brainsets.datasets.OpenNeuroDataset import OpenNeuroDataset, OpenNeuroSplitType
 
 
 class KochiVisualNamingDS006914(OpenNeuroDataset):
