+++ brainsets/datasets/ShiraziHBNR1DS005505.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional
 
-from brainsets.datasets.OpenNeuroBase import OpenNeuroDataset, OpenNeuroSplitType
+from brainsets.datasets.OpenNeuroDataset import OpenNeuroDataset, OpenNeuroSplitType
 
 
 class ShiraziHBNR1DS005505(OpenNeuroDataset):
