+++ brainsets/datasets/__init__.py
@@ -30,7 +30,7 @@ __all__ = (
     + ["OpenNeuroDataset", "OpenNeuroSplitType"]
 )
 
-from .OpenNeuroBase import OpenNeuroDataset, OpenNeuroSplitType
+from .OpenNeuroDataset import OpenNeuroDataset, OpenNeuroSplitType
 
 from .PerichMillerPopulation2018 import PerichMillerPopulation2018
 from .PeiPandarinathNLB2021 import PeiPandarinathNLB2021
